+++ changelog.md
@@ -6,6 +6,7 @@
 
 * fixed: KML maps can now use the `center` attribute to set the map center
 * added: attribute `hidefullscreen` for hiding the full-screen control
+* added: attribute `markeranimation` for setting how a single marker is added to the map - drop, bounce, none; defaults to drop
 
 ### 1.14.0, 2018-06-07